refactor federated connection into namespace
